Rename + add `typer` + refactor
