+++ 2023/openai/pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "mock-openai"
 readme = "README.md"
-version = "0.1.0"
+version = "0.2.0"
 authors = [
     { name = "Dan Miller", email = "daniel.keegan.miller@gmail.com" }
 ]
@@ -11,6 +11,7 @@ dependencies = [
     "httpx==0.23.0",
     "openai==0.27.9",
     "requests",
+    "typer==0.9.0"
 ]
 [project.optional-dependencies]
 dev = [
@@ -20,7 +21,7 @@ dev = [
 ]
 
 [project.scripts]
-mock-openai = "mock_openai.app:main"
+mock-openai = "mock_openai.app:cli"
 
 [build-system]
 requires = ["setuptools"]
